+++ src/simulation/onboardDataHandling/storageUnit/partitionedStorageUnit.cpp
@@ -70,7 +70,7 @@ void PartitionedStorageUnit::addPartition(std::string dataName){
  */
 void PartitionedStorageUnit::setDataBuffer(std::vector<std::string> partitionNames, std::vector<long long int> data){
 
-    for (int i = 0; i < partitionNames.size(); i++)
+    for (size_t i = 0; i < partitionNames.size(); i++)
     {
         PartitionedStorageUnit::DataStorageUnitBase::setDataBuffer(partitionNames[i], data[i]);
     }
